add userToken support #24
Store the dps.report userToken response and persist it in the config.
Send the stored userToken with every upload.

Add the userToken to the configuratoin UI.
This allows setting, resetting and disabling the userToken.

At the moment only one userToken can be stored but the current
implementation could be extended easily to support multiple tokens
and the ability to switch between them.